Automatic Car Parking Assistance using Fuzzy Logic
A hardware model which effectively parks the car in its assigned
position, without any human assistance using Arduino Uno and NXP’s Cup
Intelligent was implemented using this code. The shortest distance to
the destination is achieved using the A* algorithm and to achieve
parallel parking, Parallel parking algorithm is used. • The desired
speed and steering angle for the car is designed by Pulse Width
Modulation (PWM).